Fix bf16-mixed dtyping issue

+++ zatom/models/zatom.py
@@ -520,7 +520,9 @@ class Zatom(LightningModule):
                     )
                     # Apply same random translation to all (periodic) Cartesian coordinates
                     pos_aug = batch.pos + random_translation
-                    batch.pos[batch.node_is_periodic] = pos_aug[batch.node_is_periodic]
+                    batch.pos[batch.node_is_periodic] = pos_aug[batch.node_is_periodic].type(
+                        batch.pos.dtype
+                    )
                     # Compute new fractional coordinates for periodic samples
                     cell_per_node_inv = torch.linalg.inv(
                         # NOTE: `torch.linalg.inv` does not support low precision dtypes
